feat: Storage Layer Refactoring Phase 2 - User Domain Extraction (#121)
Extract user-related operations from monolithic storage.ts into UserStorage domain repository.

Files Created:
- server/storage/domains/user-storage.ts (475 lines)

Methods Extracted (15 total):
- Basic: getUserCount(), getUserByIdSafe()
- Auth: registerUser(), resetPassword(), createUserWithTransaction()
- Profile: updateUserProfile(), updateUserTrustLevel(), suspendUser()
- Admin: getAllUsers(), getAdminAnalyticsOverview(), getUserGrowthData(), getForumActivityData(), getTopCategories()

Security: NEVER exposes passwordHash (SafeUser type exclusively, all code lines have security markers)

Statistics:
- Before: storage.ts 7,035 lines
- After: storage.ts ~6,535 lines, user-storage.ts 475 lines
- Methods extracted: 15/86 (~17%)

Related: #121
